Review feedback

+++ features/events/daemon/src/main/java/org/opennms/netmgt/eventd/EventUtil.java
@@ -70,7 +70,7 @@ public interface EventUtil {
 	 */
 	String getIfAlias(long nodeId, String ipAddr) throws SQLException;
 
-	String getIfAliasByNodeAndIfIndex(long nodeId, long ifIndex) throws SQLException;
+	String getIfAliasByNodeAndIfIndex(long nodeId, int ifIndex) throws SQLException;
 
 	/**
 	 * Helper method.
